+++ ui/src/pages/Documents/Documents.module.scss
@@ -30,3 +30,35 @@
 .filedivider {
     border-bottom: 2px solid $foreground1;
 }
+
+.upload {
+    display: flex;
+
+    flex: 1;
+    align-items: center;
+    flex-direction: column;
+
+    margin: 0;
+    padding: 20px;
+    border: 2px dashed;
+    border-top: none;
+    color: $foreground1;
+    background-color: $background1;
+
+    outline: none;
+    transition: border .24s ease-in-out;
+
+    border-color: $foreground1;
+
+    &.uploadActive {
+        border-color: $action;
+    }
+
+    &.uploadAccept {
+        border-color: $accept;
+    }
+
+    &.uploadReject {
+        border-color: $reject;
+    }
+}
