--- ui/src/pages/Home/Home.module.scss
@@ -1,33 +1,0 @@
-@import "../../theme.scss";
-
-.base {
-    display: flex;
-
-    flex: 1;
-    align-items: center;
-    flex-direction: column;
-
-    margin: 0;
-    padding: 20px;
-    border: 2px dashed;
-    border-top: none;
-    color: $foreground1;
-    background-color: $background1;
-
-    outline: none;
-    transition: border .24s ease-in-out;
-
-    border-color: $foreground1;
-
-    &.active {
-        border-color: $action;
-    }
-
-    &.accept {
-        border-color: $accept;
-    }
-
-    &.reject {
-        border-color: $reject;
-    }
-}
